Move all types to a single file
