+++ lib/postgrest.dart
@@ -1,11 +1,7 @@
-export 'src/count_option.dart';
 export 'src/postgrest.dart';
 export 'src/postgrest_builder.dart';
-export 'src/postgrest_error.dart';
 export 'src/postgrest_filter_builder.dart';
 export 'src/postgrest_query_builder.dart';
-export 'src/postgrest_response.dart';
 export 'src/postgrest_rpc_builder.dart';
 export 'src/postgrest_transform_builder.dart';
-export 'src/returning_option.dart';
-export 'src/text_search_type.dart';
+export 'src/types.dart';
