--- lib/src/returning_option.dart
@@ -1,11 +1,0 @@
-/// Returns count as part of the response when specified.
-enum ReturningOption {
-  minimal,
-  representation,
-}
-
-extension ReturningOptionName on ReturningOption {
-  String name() {
-    return toString().split('.').last;
-  }
-}
