--- lib/src/text_search_type.dart
@@ -1,18 +1,0 @@
-/// The type of tsquery conversion to use on [query].
-enum TextSearchType {
-  /// Uses PostgreSQL's plainto_tsquery function.
-  plain,
-
-  /// Uses PostgreSQL's phraseto_tsquery function.
-  phrase,
-
-  /// Uses PostgreSQL's websearch_to_tsquery function.
-  /// This function will never raise syntax errors, which makes it possible to use raw user-supplied input for search, and can be used with advanced operators.
-  websearch,
-}
-
-extension TextSearchTypeName on TextSearchType {
-  String name() {
-    return toString().split('.').last;
-  }
-}
